+++ lib/ui/page/add_task/form/add_task_form_controller.dart
@@ -1,12 +1,12 @@
 import 'package:formz/formz.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
-import 'package:tokidoki_mobile/ui/common_validation/task_form.dart';
-import 'package:tokidoki_mobile/ui/common_validation/task_validation.dart';
+import 'package:tokidoki_mobile/ui/common_form/task_form.dart';
+import 'package:tokidoki_mobile/ui/common_form/task_validation.dart';
 
-part 'edit_task_form_controller.g.dart';
+part 'add_task_form_controller.g.dart';
 
 @riverpod
-class EditTaskFormController extends _$EditTaskFormController {
+class AddTaskFormController extends _$AddTaskFormController {
   @override
   TaskForm build() {
     return TaskForm(
